Capstone 1/app/requirements.txt
numpy==1.24.0
Pillow==9.3.0
fastapi==0.88.0
uvicorn==0.20.0
python-multipart==0.0.5